Fix user_id FK type in trainee_document_comments migration
Match users.id which uses increments() (unsignedInteger),
not unsignedBigInteger — prevents SQLSTATE[HY000] 3780 on MySQL 8.

+++ database/migrations/2026_05_18_200001_create_trainee_document_comments_table.php
@@ -10,7 +10,7 @@ class CreateTraineeDocumentCommentsTable extends Migration
         Schema::create('trainee_document_comments', function (Blueprint $table) {
             $table->increments('id');
             $table->unsignedInteger('trainee_document_id');
-            $table->unsignedBigInteger('user_id');
+            $table->unsignedInteger('user_id');
             $table->text('comment');
             $table->timestamps();
             $table->foreign('trainee_document_id')->references('id')->on('trainee_documents')->onDelete('cascade');
